requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
openpyxl==3.1.5
PyMuPDF==1.24.10
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
aiosqlite==0.20.0
